Add somes features

+++ src/events/messageCreate.event.ts
@@ -1,12 +1,16 @@
 import { Client, Message, PermissionFlagsBits } from "discord.js";
+import { Logging } from "../utils/logging.util";
+import { Configuration } from "../utils/config.util";
 import { IEvent } from "../utils/interfaces/event.interface";
+import { EmbedsUtil } from "../utils/embeds.util";
 
 const event: IEvent = {
     name: 'messageCreate',
     execute(app: Client, message: Message){
-        if(message.channelId === '988381575018348556'){
+
+        if(new Configuration().getCommandsOnly(message.guildId || '').includes(message.channelId)){
             if(!message.member?.permissions.has(PermissionFlagsBits.Administrator)){
-                message.delete();
+                Logging.loggingChannel?.send({embeds: [EmbedsUtil.info("🖊️ Blocked message", [`${message.author} try to send \`${message.content}\` in ${message.channel}`])]}).then(() => message.delete());
             }
         }
     }
